requirements.txt
amqp==5.2.0
APScheduler==3.10.4
asgiref==3.8.1
billiard==4.2.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.0.6
django-humanize==0.1.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
humanize==4.9.0
kombu==5.3.7
prompt_toolkit==3.0.47
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
redis==5.0.6
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
tzlocal==5.2
vine==5.1.0
wcwidth==0.2.13
